moved to python folder and state management
