+++ python/.gitignore
@@ -1,0 +1,16 @@
+venv/
+output/*
+samples/
+other/
+models/other/
+log*
+TEST \-*
+__pycache__/
+.vscode/
+archive
+demo
+intermediate_photos/
+hardware_accel_interfacing/
+.idea*
+debug_tools/custom_data/test/
+debug_tools/custom_data/train/
