+++ python/requirements/requirements.txt
@@ -1,0 +1,2 @@
+opencv-python==4.5.1.48
+imutils==0.5.4
