+++ python/requirements/requirements_for_tensorflow.txt
@@ -1,0 +1,2 @@
+tensorflow==2.0.0a0
+h5py==2.10.0
