--- requirements/requirements.txt
@@ -1,2 +1,0 @@
-opencv-python==4.5.1.48
-imutils==0.5.4
