--- requirements/requirements_for_tensorflow.txt
@@ -1,2 +1,0 @@
-tensorflow==2.0.0a0
-h5py==2.10.0
